# backend/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Bookstore API
  version: 1.0.0
  description: API documentation for the online e-book store. Manages users, authentication, books, reviews, cart, orders, and payments.
servers:
  - url: http://18.184.165.152:3000
    description: API Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  # 2. Define Reusable Schemas (from Models)
  schemas:
...
    Error404:
      type: object
      properties:
        error:
          type: string
          example: User not found
...
  responses:
...
    Error404:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error404'
...
paths:
...
  /api/payments/create-checkout/{orderId}:
    post:
      tags: [Payments]
      summary: (User) Create Stripe Checkout Session
      description: Creates a Stripe Checkout Session for a 'pending' order.
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkoutUrl:
                    type: string
                    format: uri
                    example: https://checkout.stripe.com/c/session_...
                  sessionId:
                    type: string
                    example: cs_test_abc123
        '400':
          description: Order is not pending or currency mismatch
        '403':
          description: Forbidden (not order owner)
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          description: Failed to create checkout session
  /api/payments/test-confirm/{orderId}:
    post:
      tags: [Payments]
      summary: (User - TEST) Confirm a payment for testing
      description: Bypasses Stripe checkout for testing; directly confirms payment.
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_method:
                  type: string
                  example: pm_card_visa
      responses:
        '200':
          description: Payment succeeded & order marked as paid
        '400':
          description: Payment not succeeded
          
  # --- Upload ---
...